Update users variable to match rest of galaxy
There are many roles in ansible galaxy which are using users variable.
Most of these roles expect users to be a dictionary rather than a
simple list, with a 'username' entry for each user (along with other
options). In order for this role to be compatible with other roles,
update the users variable accordingly.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@email.com>